Extract PluginTask class to independent class file
